Rename: 채널과 메시지 패키지 분리
