+++ src/main/java/com/vigilante/retriever/v1/message/application/query/MessageMongoQuery.java
@@ -1,13 +1,13 @@
-package com.vigilante.retriever.v1.channel.application.query;
+package com.vigilante.retriever.v1.message.application.query;
 
-import static com.vigilante.retriever.v1.channel.domain.code.MessageErrorCode.*;
+import static com.vigilante.retriever.v1.message.domain.code.MessageErrorCode.*;
 
 import java.util.List;
 
 import com.vigilante.retriever.common.domain.annotation.QueryService;
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
-import com.vigilante.retriever.v1.channel.domain.entity.MessageEntity;
-import com.vigilante.retriever.v1.channel.domain.port.out.MessageMongoPort;
+import com.vigilante.retriever.v1.message.domain.entity.MessageEntity;
+import com.vigilante.retriever.v1.message.domain.port.out.MessageMongoPort;
 
 import lombok.RequiredArgsConstructor;
 
@@ -22,7 +22,8 @@ public class MessageMongoQuery {
 	}
 
 	public MessageEntity getById(String id) {
-		return messageMongoPort.findById(id).orElseThrow(() -> new NotFoundException(MESSAGE_NOT_FOUND));
+		return messageMongoPort.findById(id)
+			.orElseThrow(() -> new NotFoundException(MESSAGE_NOT_FOUND));
 	}
 
 	public List<MessageEntity> findByChannelId(long channelId) {
