+++ src/main/java/com/vigilante/retriever/v1/message/domain/entity/MessageEntity.java
@@ -1,9 +1,9 @@
-package com.vigilante.retriever.v1.channel.domain.entity;
+package com.vigilante.retriever.v1.message.domain.entity;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.vigilante.retriever.v1.channel.domain.enums.SenderType;
+import com.vigilante.retriever.v1.message.domain.enums.SenderType;
 
 import lombok.Builder;
